Rewrite fuzzy-phrase bindings to support neon 0.7
